Extend exporter ACL policy
This allows the exporter to view details about a job which is relevant
for the content of the `project/.../executions` endpoint. Without this
permission the response never contains details about executions so all
the exporter knows about is running executions but not what happened
with the executions which makes it a bit pointless.